Move props to components
