+++ components/basecomponents/KidsDictionaryList.tsx
@@ -4,7 +4,11 @@ import KidsTranslationContainer from './KidsTranslationContainer';
 /* Hooks, Types, Utils */
 import { normalizeForId } from 'utils/textNormalizationUtils';
 import { getCountryVariant } from 'utils/locales';
-import { KidsCategoryListProps } from '@types';
+import { Category } from 'utils/getDataUtils';
+
+export type KidsCategoryListProps = {
+  kidsCategory: Category | undefined;
+};
 
 const KidsDictionaryList: React.FC<KidsCategoryListProps> = ({ kidsCategory }) => {
   if (!kidsCategory?.translations) {
